pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robot_framework"
version = "0.0.2"
authors = [
  { name="MBU", email="rpa@mbu.aarhus.dk" },
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "OpenOrchestrator == 1.*",
    "Pillow == 10.*",
    "requests_ntlm >= 1.2.0"
]

[project.optional-dependencies]
dev = [
  "pylint",
  "flake8"
]
